Add ui

+++ app/src/main/java/com/rijksmuseum/sample/models/Price.kt
@@ -4,7 +4,7 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class Price(
-    val id: String,
-    val calculationType: Int,
-    val amount: Int
+        val id: String,
+        val calculationType: Int,
+        val amount: Double
 )
